network_components: wait for lnd to finish rescan
This commit adds a way to observe the output logs of lnd nodes
and block the startup of new nodes by waiting a certain time.
The time it takes to start a node is determined by the log output
"Finished rescan", which is parsed for. This obsoletes the sleep
timer, which was unreliable.